[MOD] Static functions aren't expressions.
