+++ src/main/java/org/basex/query/func/InlineFunc.java
@@ -1,6 +1,7 @@
 package org.basex.query.func;
 
 import static org.basex.query.QueryText.*;
+import static org.basex.query.util.Err.*;
 
 import java.util.*;
 import java.util.Map.Entry;
@@ -24,7 +25,27 @@ import org.basex.util.list.*;
  * @author BaseX Team 2005-12, BSD License
  * @author Leo Woerteler
  */
-public final class InlineFunc extends UserFunc {
+public final class InlineFunc extends Single implements Scope {
+  /** Function name. */
+  private final QNm name;
+  /** Arguments. */
+  private final Var[] args;
+  /** Return type. */
+  private final SeqType ret;
+  /** Annotations. */
+  private final Ann ann;
+  /** Updating flag. */
+  private final boolean updating;
+
+  /** Map with requested function properties. */
+  private final EnumMap<Use, Boolean> map = new EnumMap<Expr.Use, Boolean>(Use.class);
+  /** Static context. */
+  private final StaticContext sc;
+  /** Compilation flag. */
+  private boolean compiled;
+
+  /** Local variables in the scope of this function. */
+  private final VarScope scope;
   /**
    * Constructor.
    * @param ii input info
@@ -53,28 +74,134 @@ public final class InlineFunc extends UserFunc {
    */
   InlineFunc(final InputInfo ii, final QNm nm, final SeqType r, final Var[] v,
       final Expr e, final Ann a, final StaticContext stc, final VarScope scp) {
-    super(ii, nm, v, r, a, e, stc, scp);
-    expr = e;
+    super(ii, e);
+    name = nm;
+    args = v;
+    ret = r;
+    ann = a == null ? new Ann() : a;
+    updating = ann.contains(Ann.Q_UPDATING);
+    scope = scp;
+    sc = stc;
+  }
+
+  @Override
+  public void compile(final QueryContext ctx) throws QueryException {
+    compile(ctx, null);
+  }
+
+  /**
+   * Removes and returns all static bindings from this function's closure.
+   * @return static variable bindings
+   */
+  private Collection<Entry<Var, Expr>> staticBindings() {
+    Collection<Entry<Var, Expr>> propagate = null;
+    final Iterator<Entry<Var, Expr>> cls = scope.closure().entrySet().iterator();
+    while(cls.hasNext()) {
+      final Entry<Var, Expr> e = cls.next();
+      final Expr c = e.getValue();
+      if(c.isValue()) {
+        if(propagate == null) propagate = new ArrayList<Entry<Var, Expr>>();
+        propagate.add(e);
+        cls.remove();
+      }
+    }
+    return propagate == null ? Collections.<Entry<Var, Expr>>emptyList() : propagate;
   }
 
   @Override
   public Expr compile(final QueryContext ctx, final VarScope scp)
       throws QueryException {
-    comp(ctx, scp);
+    if(compiled) return this;
+    compiled = true;
+
+    // compile closure
+    for(final Entry<Var, Expr> e : scope.closure().entrySet())
+      e.setValue(e.getValue().compile(ctx, scp));
+
+    final StaticContext tmp = ctx.sc;
+    ctx.sc = sc;
+
+    final int fp = scope.enter(ctx);
+    try {
+      // constant propagation
+      for(final Entry<Var, Expr> e : staticBindings())
+        ctx.set(e.getKey(), e.getValue().value(ctx), info);
+
+      expr = expr.compile(ctx, scope);
+    } catch(final QueryException qe) {
+      expr = FNInfo.error(qe, info);
+    } finally {
+      scope.cleanUp(this);
+      scope.exit(ctx, fp);
+      ctx.sc = tmp;
+    }
+
     return optimize(ctx, scp);
   }
 
   @Override
   public Expr optimize(final QueryContext ctx, final VarScope scp) throws QueryException {
-    type = FuncType.get(this).seqType();
+    type = FuncType.get(args, ret).seqType();
     size = 1;
     // only evaluate if the closure is empty, so we don't lose variables
     return scope.closure().isEmpty() ? preEval(ctx) : this;
   }
 
   @Override
+  public VarUsage count(final Var v) {
+    VarUsage all = VarUsage.NEVER;
+    for(final Entry<Var, Expr> e : scope.closure().entrySet())
+      if((all = all.plus(e.getValue().count(v))) == VarUsage.MORE_THAN_ONCE) break;
+    return all;
+  }
+
+  @Override
+  public Expr inline(final QueryContext ctx, final VarScope scp,
+      final Var v, final Expr e) throws QueryException {
+    boolean change = false, val = false;
+
+    for(final Entry<Var, Expr> entry : scope.closure().entrySet()) {
+      final Expr ex = entry.getValue().inline(ctx, scp, v, e);
+      if(ex != null) {
+        change = true;
+        val |= ex.isValue();
+        entry.setValue(ex);
+      }
+    }
+
+    final StaticContext tmp = ctx.sc;
+    ctx.sc = sc;
+
+    if(val) {
+      final int fp = scope.enter(ctx);
+      try {
+        for(final Entry<Var, Expr> entry : staticBindings()) {
+          final Expr inl = expr.inline(ctx, scope, entry.getKey(), entry.getValue());
+          if(inl != null) expr = inl;
+        }
+      } catch(final QueryException qe) {
+        expr = FNInfo.error(qe, info);
+      } finally {
+        scope.cleanUp(this);
+        scope.exit(ctx, fp);
+        ctx.sc = tmp;
+      }
+    }
+
+    return change ? optimize(ctx, scp) : null;
+  }
+
+  @Override
+  public Expr copy(final QueryContext cx, final VarScope scp, final IntMap<Var> vs) {
+    final VarScope v = scope.copy(cx, scp, vs);
+    final Var[] a = args.clone();
+    for(int i = 0; i < a.length; i++) a[i] = vs.get(a[i].id);
+    return copyType(new InlineFunc(info, name, ret, a, expr.copy(cx, v, vs), ann, sc, v));
+  }
+
+  @Override
   public FuncItem item(final QueryContext ctx, final InputInfo ii) throws QueryException {
-    final FuncType ft = FuncType.get(this);
+    final FuncType ft = FuncType.get(args, ret);
     final boolean c = ft.ret != null && !expr.type().instanceOf(ft.ret);
 
     // collect closure
@@ -97,21 +224,27 @@ public final class InlineFunc extends UserFunc {
 
   @Override
   public boolean uses(final Use u) {
-    return u == Use.X30 || super.uses(u);
+    if(u == Use.X30) return true;
+
+    // handle recursive calls: set dummy value, eventually replace it with final value
+    Boolean b = map.get(u);
+    if(b == null) {
+      map.put(u, false);
+      b = expr == null || super.uses(u);
+      map.put(u, b);
+    }
+    return b;
   }
 
   @Override
   public boolean removable(final Var v) {
+    // [LW] Variables are removable from the closure.
     return false;
   }
 
   @Override
   public boolean databases(final StringList db) {
-    return false;
-  }
-
-  @Override
-  protected boolean tco() {
+    // [LW][JE] can be done better...
     return false;
   }
 
@@ -125,20 +258,12 @@ public final class InlineFunc extends UserFunc {
   }
 
   @Override
-  public Expr copy(final QueryContext cx, final VarScope scp, final IntMap<Var> vs) {
-    final VarScope v = scope.copy(cx, scp, vs);
-    final Var[] a = args.clone();
-    for(int i = 0; i < a.length; i++) a[i] = vs.get(a[i].id);
-    return copyType(new InlineFunc(info, name, ret, a, expr.copy(cx, v, vs), ann, sc, v));
-  }
-
-  @Override
   public boolean visit(final ASTVisitor visitor) {
     final Map<Var, Expr> clos = scope.closure();
     for(final Entry<Var, Expr> v : clos.entrySet())
       if(!(v.getValue().accept(visitor) && visitor.declared(v.getKey()))) return false;
     for(final Var v : args) if(!visitor.declared(v)) return false;
-    return visitAll(visitor, expr);
+    return expr.accept(visitor);
   }
 
   @Override
@@ -151,5 +276,49 @@ public final class InlineFunc extends UserFunc {
     sb.append(PAR2).append(' ');
     if(ret != null) sb.append("as ").append(ret.toString()).append(' ');
     return sb.append("{ ").append(expr).append(" }").toString();
+  }
+
+  @Override
+  public void checkUp() throws QueryException {
+    final boolean u = expr.uses(Use.UPD);
+    if(u) expr.checkUp();
+    if(updating) {
+      // updating function
+      if(ret != null) UPFUNCTYPE.thrw(info);
+      if(!u && !expr.isVacuous()) UPEXPECTF.thrw(info);
+    } else if(u) {
+      // uses updates, but is not declared as such
+      UPNOT.thrw(info, description());
+    }
+  }
+
+  @Override
+  public boolean isVacuous() {
+    return !uses(Use.UPD) && ret != null && ret.eq(SeqType.EMP);
+  }
+
+  @Override
+  public SeqType type() {
+    return updating ? SeqType.EMP : super.type();
+  }
+
+  @Override
+  public boolean accept(final ASTVisitor visitor) {
+    for(final Entry<Var, Expr> e : scope.closure().entrySet())
+      if(!e.getValue().accept(visitor)) return false;
+    return visitor.inlineFunc(this);
+  }
+
+  @Override
+  public int exprSize() {
+    int sz = 1;
+    for(final Entry<Var, Expr> e : scope.closure().entrySet())
+      sz += e.getValue().exprSize();
+    return sz + expr.exprSize();
+  }
+
+  @Override
+  public boolean compiled() {
+    return compiled;
   }
 }
